+++ tket-py/tket/extensions/__init__.py
@@ -12,9 +12,17 @@ from tket_exts import (
     wasm,
 )
 
+# TODO: Remove once tket no longer supports tket-exts 0.10.*
+try:
+    from tket_exts import bool  # type: ignore[attr-defined]
+except ImportError:
+    bool = opaque_bool  # type: ignore[assignment]
 
+
+# TODO: Remove the deprecated `opaque_bool` export in a breaking change.
 __all__ = [
     "debug",
+    "bool",
     "opaque_bool",
     "guppy",
     "rotation",