+++ tket-py/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [0.12.4](https://github.com/CQCL/tket2/compare/tket-py-v0.12.3...tket-py-v0.12.4) (2025-08-25)
+
+
+### Bug Fixes
+
+* Pytket barrier operations not being decoded ([#1069](https://github.com/CQCL/tket2/issues/1069)) ([4b90ffd](https://github.com/CQCL/tket2/commit/4b90ffdc7de6ec696d5db5e946b417bf8ff71878))
+
+## [0.12.3](https://github.com/CQCL/tket2/compare/tket-py-v0.12.2...tket-py-v0.12.3) (2025-08-22)
+
+
+### Features
+
+* Explicit exports for tket_exts ops and types ([#1046](https://github.com/CQCL/tket2/issues/1046)) ([a32873e](https://github.com/CQCL/tket2/commit/a32873e3543b7d77f3bed08016485ab292f5204a))
+
+
+### Bug Fixes
+
+* Fix erroneous parameters being decoded from pytket for qsystem gates ([#1061](https://github.com/CQCL/tket2/issues/1061)) ([cd42644](https://github.com/CQCL/tket2/commit/cd42644ccd50a60795e527efdfbe0727344db373))
+
 ## [0.12.2](https://github.com/CQCL/tket2/compare/tket-py-v0.12.1...tket-py-v0.12.2) (2025-08-19)